Move single element tuple like traits
