+++ include/iris/x4/traits/tuple_traits.hpp
@@ -67,7 +67,31 @@ struct unwrap_single_element_tuple_like<T>
     using type = alloy::tuple_element_t<0, T>;
 };
 
+template<class T>
+struct unwrap_single_element_plain
+{
+    using type = std::remove_cvref_t<T>;
+};
 
+template<class T>
+    requires traits::is_single_element_tuple_like<std::remove_cvref_t<T>>::value
+struct unwrap_single_element_plain<T>
+{
+    using type = std::remove_cvref_t<alloy::tuple_element_t<0, T>>;
+};
+
+template<class T>
+[[nodiscard]] constexpr auto&& unwrap_single_element(T&& value) noexcept
+{
+    return std::forward<T>(value);
+}
+
+template<class T>
+    requires traits::is_single_element_tuple_like<std::remove_cvref_t<T>>::value
+[[nodiscard]] constexpr auto&& unwrap_single_element(T&& value) noexcept
+{
+    return std::forward_like<T>(alloy::get<0>(std::forward<T>(value)));
+}
 
 } // iris::x4::traits
